backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.6
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
iniconfig==1.1.1
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
